調整對齊的介面，予程式較好維護，models莫閣函式較清楚
