+++ 臺灣言語服務/Kaldi語料對齊.py
@@ -14,7 +14,9 @@ from 臺灣言語工具.系統整合.程式腳本 import 程式腳本
 from 臺灣言語服務.models import Kaldi對齊結果
 
 
-class Kaldi語料對齊:
+class Kaldi語料對齊(Kaldi對齊結果):
+    class Meta:
+        proxy = True
 
     @classmethod
     def 匯入音檔(cls, 語言, 啥人唸的, 聲音檔, 內容):
@@ -41,19 +43,21 @@ class Kaldi語料對齊:
         ]
         聽拍內容 = {'聽拍資料': 聽拍資料}
         聽拍內容.update(公家內容)
-        聽拍=影音.寫聽拍(聽拍內容)
-        
-        Kaldi對齊結果.準備對齊(影音, 聽拍)
-        return 影音
+        聽拍 = 影音.寫聽拍(聽拍內容)
+
+        return cls._準備對齊(影音, 聽拍)
 
     @classmethod
-    def 對齊音檔(cls, 影音):
-        語言 = 影音.語言腔口.語言腔口
+    def _準備對齊(cls, 影音, 聽拍):
+        return cls.objects.create(影音=影音, 欲切開的聽拍=聽拍)
+
+    def 對齊音檔(self):
+        語言 = self.影音.語言腔口.語言腔口
         服務設定 = settings.HOK8_BU7_SIAT4_TING7[語言]
 
         對齊設定 = 服務設定['辨識設定']
         kaldi_eg目錄 = 對齊設定['腳本資料夾']
-        影音編號 = 影音.編號()
+        影音編號 = self.影音.編號()
 
         編號字串資料夾名 = '{0:07}'.format(影音編號)
         暫存目錄 = join(settings.BASE_DIR, 'kaldi資料')
@@ -86,3 +90,43 @@ class Kaldi語料對齊:
                 {'開始': float(開始時間), '長度': float(長度), '分詞': 分詞}
             )
         return 對齊文本
+
+    def 對齊成功(self, ctm時間):
+        self.對齊好猶未 = True
+        self.對齊出問題 = False
+        self.save()
+
+        公家內容 = {
+            '收錄者': 來源表.objects.get_or_create(名='系統管理員')[0].編號(),
+            '來源': 來源表.objects.get_or_create(名='系統管理員')[0].編號(),
+            '版權': 版權表.objects.get_or_create(版權='會使公開')[0].pk,
+            '種類': '語句',
+            '語言腔口': self.欲切開的聽拍.語言腔口.語言腔口,
+            '著作所在地': '臺灣',
+            '著作年': str(timezone.now().year),
+        }
+
+        聽拍資料 = [
+        ]
+        ctm所在 = 0
+        for 一段 in self.欲切開的聽拍.聽拍內容()[0]['內容'].split('\n'):
+            這段長度 = len(一段.split())
+            這段資訊 = ctm時間[ctm所在:ctm所在 + 這段長度]
+            聽拍資料.append(
+                {
+                    '語者': '媠媠',
+                    '內容': 一段,
+                    '開始時間': 這段資訊[0]['開始'],
+                    '結束時間': 這段資訊[-1]['開始'] + 這段資訊[-1]['長度'],
+                }
+            )
+            ctm所在 = +這段長度
+        聽拍內容 = {'聽拍資料': 聽拍資料}
+        聽拍內容.update(公家內容)
+        self.切好的聽拍 = self.欲切開的聽拍.校對做(聽拍內容)
+        self.save()
+
+    def 對齊失敗(self):
+        self.對齊好猶未 = True
+        self.對齊出問題 = True
+        self.save()
